codereport/requirements.txt
jinja2
python-slugify
Pygments
